ff_unpack.c: 1. Updated column_attributes to use ints (instead of shorts) for bytesize and adj_bytesize. Some repetition count values were overflowing shorts.
ff_interp.c:
1. Added nearest and nearest_before options to ff_interp
2. Updated value comparison from floats to doubles

ff_ascii.c:
1. Added a TODO about handling null characters in strings

docs/dv.gih:
1. Updated documentation for interp()



git-svn-id: http://oss.mars.asu.edu/svn/davinci/davinci/trunk@19732 dedd6328-c50f-0410-8be5-b54a49638459

+++ ff_unpack.c
@@ -50,9 +50,9 @@ typedef unsigned char byte;
 typedef struct
 {
 	char type;					//type constant from above defines
-	short bytesize;				//bytesize in file (from template)
-	short adj_bytesize;			//bytesize allocated (e.g. +1 for strings for null character
-	short columns;				//multiplicity
+	int bytesize;				//bytesize in file (from template)
+	int adj_bytesize;			//bytesize allocated (e.g. +1 for strings for null character
+	int columns;				//multiplicity
 	unsigned short start_byte;	//record offset
 	char* col_name;				//generated name used for creating davinci structs
 } column_attributes;
